requirements.txt
# LangChain Core
langchain
langchain-core

# OpenAI integration
langchain-openai
openai

# Anthropic integration
langchain-anthropic

# Google Gemini Integration
langchain-google-genai
google-generativeai

# HuggingFace integration
langchain-huggingface
transformers
huggingface-hub

# Environment variable management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn